net: u64 length compare in write_bytes (wasm-safe)
write_bytes asserted the byte length fits in u32 via int(max(u32)) —
which overflows a 32-bit int at compile time on wasm32. Scrapyard is the
first kit CONSUMER to compile the kit to WebAssembly (coop_arena's web
build uses Godot RPCs, not the kit), so this portability bug only just
surfaced. Compare in u64 instead: correct on both int widths, and the
guard still holds. Suite ALL GREEN; the kit now compiles clean to wasm.

+++ kit/net/wire.odin
@@ -89,7 +89,7 @@ write_string :: proc(w: ^Writer, s: string) {
 
 write_bytes :: proc(w: ^Writer, b: []u8) {
 	n := len(b)
-	assert(n <= int(max(u32)))
+	assert(u64(n) <= u64(max(u32))) // u64 compare: `int(max(u32))` overflows a 32-bit int (wasm)
 	write_u32(w, u32(n))
 	if n > 0 {
 		write_raw(w, raw_data(b), n)
